backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
web3==6.11.3
eth-account==0.10.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.2
shap==0.43.0
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.4
sqlalchemy==2.0.23
aiofiles==23.2.1
pydantic==2.5.0
ipfshttpclient==0.8.0a2
pytest==7.4.3
pytest-asyncio==0.21.1
